backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.0.1
